+++ CHANGELOG.md
@@ -14,6 +14,9 @@ Prefix the change with one of these keywords:
 - _Fixed_: for any bug fixes.
 - _Security_: in case of vulnerabilities.
 
+## Unreleased
+- Added: useEvents and useMapEvents hook
+
 ## 0.4.0
 
 - Changed: removed L (leaflet) from mapInstance, export mapInstance directly